Método listarInscricoesNoEvento movido para InscricaoEventoAdater
